Extract SslContext configuration into SslContextFactory
