+++ spring-cloud-gateway-server/src/main/java/org/springframework/cloud/gateway/config/HttpClientFactory.java
@@ -16,25 +16,11 @@
 
 package org.springframework.cloud.gateway.config;
 
-import java.io.IOException;
-import java.net.URL;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.NoSuchProviderException;
-import java.security.cert.Certificate;
-import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.TrustManagerFactory;
-
 import io.netty.channel.ChannelOption;
-import io.netty.handler.ssl.SslContextBuilder;
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
 import reactor.netty.http.Http11SslContextSpec;
 import reactor.netty.http.Http2SslContextSpec;
@@ -50,7 +36,6 @@ import org.springframework.boot.autoconfigure.web.ServerProperties;
 import org.springframework.boot.context.properties.PropertyMapper;
 import org.springframework.core.annotation.AnnotationAwareOrderComparator;
 import org.springframework.util.CollectionUtils;
-import org.springframework.util.ResourceUtils;
 import org.springframework.util.StringUtils;
 
 import static org.springframework.cloud.gateway.config.HttpClientProperties.Pool.PoolType.DISABLED;
@@ -69,12 +54,15 @@ public class HttpClientFactory extends AbstractFactoryBean<HttpClient> {
 
 	protected final ServerProperties serverProperties;
 
+	protected final SslContextFactory sslContextFactory;
+
 	protected final List<HttpClientCustomizer> customizers;
 
 	public HttpClientFactory(HttpClientProperties properties, ServerProperties serverProperties,
-			List<HttpClientCustomizer> customizers) {
+			SslContextFactory sslContextFactory, List<HttpClientCustomizer> customizers) {
 		this.properties = properties;
 		this.serverProperties = serverProperties;
+		this.sslContextFactory = sslContextFactory;
 		this.customizers = customizers;
 	}
 
@@ -102,7 +90,7 @@ public class HttpClientFactory extends AbstractFactoryBean<HttpClient> {
 
 		httpClient = configureProxy(httpClient);
 
-		httpClient = configureSsl(httpClient);
+		httpClient = sslContextFactory.configureSsl(httpClient);
 
 		if (properties.isWiretap()) {
 			httpClient = httpClient.wiretap(true);
@@ -125,43 +113,6 @@ public class HttpClientFactory extends AbstractFactoryBean<HttpClient> {
 			}
 		}
 		return httpClient;
-	}
-
-	protected HttpClient configureSsl(HttpClient httpClient) {
-		HttpClientProperties.Ssl ssl = properties.getSsl();
-		if ((ssl.getKeyStore() != null && ssl.getKeyStore().length() > 0)
-				|| getTrustedX509CertificatesForTrustManager().length > 0 || ssl.isUseInsecureTrustManager()) {
-			httpClient = httpClient.secure(sslContextSpec -> {
-				// configure ssl
-				configureSslContext(ssl, sslContextSpec);
-			});
-		}
-		return httpClient;
-	}
-
-	protected void configureSslContext(HttpClientProperties.Ssl ssl, SslProvider.SslContextSpec sslContextSpec) {
-		SslProvider.ProtocolSslContextSpec clientSslContext = (serverProperties.getHttp2().isEnabled())
-				? Http2SslContextSpec.forClient() : Http11SslContextSpec.forClient();
-		clientSslContext.configure(sslContextBuilder -> {
-			X509Certificate[] trustedX509Certificates = getTrustedX509CertificatesForTrustManager();
-			if (trustedX509Certificates.length > 0) {
-				setTrustManager(sslContextBuilder, trustedX509Certificates);
-			}
-			else if (ssl.isUseInsecureTrustManager()) {
-				setTrustManager(sslContextBuilder, InsecureTrustManagerFactory.INSTANCE);
-			}
-
-			try {
-				sslContextBuilder.keyManager(getKeyManagerFactory());
-			}
-			catch (Exception e) {
-				logger.error(e);
-			}
-		});
-
-		sslContextSpec.sslContext(clientSslContext).handshakeTimeout(ssl.getHandshakeTimeout())
-				.closeNotifyFlushTimeout(ssl.getCloseNotifyFlushTimeout())
-				.closeNotifyReadTimeout(ssl.getCloseNotifyReadTimeout());
 	}
 
 	protected HttpClient configureProxy(HttpClient httpClient) {
@@ -201,82 +152,6 @@ public class HttpClientFactory extends AbstractFactoryBean<HttpClient> {
 		return spec;
 	}
 
-	protected X509Certificate[] getTrustedX509CertificatesForTrustManager() {
-		HttpClientProperties.Ssl ssl = properties.getSsl();
-
-		try {
-			CertificateFactory certificateFactory = CertificateFactory.getInstance("X.509");
-			ArrayList<Certificate> allCerts = new ArrayList<>();
-			for (String trustedCert : ssl.getTrustedX509Certificates()) {
-				try {
-					URL url = ResourceUtils.getURL(trustedCert);
-					Collection<? extends Certificate> certs = certificateFactory.generateCertificates(url.openStream());
-					allCerts.addAll(certs);
-				}
-				catch (IOException e) {
-					throw new RuntimeException("Could not load certificate '" + trustedCert + "'", e);
-				}
-			}
-			return allCerts.toArray(new X509Certificate[allCerts.size()]);
-		}
-		catch (CertificateException e1) {
-			throw new RuntimeException("Could not load CertificateFactory X.509", e1);
-		}
-	}
-
-	protected KeyManagerFactory getKeyManagerFactory() {
-		HttpClientProperties.Ssl ssl = properties.getSsl();
-		try {
-			if (ssl.getKeyStore() != null && ssl.getKeyStore().length() > 0) {
-				KeyManagerFactory keyManagerFactory = KeyManagerFactory
-						.getInstance(KeyManagerFactory.getDefaultAlgorithm());
-				char[] keyPassword = ssl.getKeyPassword() != null ? ssl.getKeyPassword().toCharArray() : null;
-
-				if (keyPassword == null && ssl.getKeyStorePassword() != null) {
-					keyPassword = ssl.getKeyStorePassword().toCharArray();
-				}
-
-				keyManagerFactory.init(this.createKeyStore(), keyPassword);
-
-				return keyManagerFactory;
-			}
-
-			return null;
-		}
-		catch (Exception e) {
-			throw new IllegalStateException(e);
-		}
-	}
-
-	protected KeyStore createKeyStore() {
-		HttpClientProperties.Ssl ssl = properties.getSsl();
-		try {
-			KeyStore store = ssl.getKeyStoreProvider() != null
-					? KeyStore.getInstance(ssl.getKeyStoreType(), ssl.getKeyStoreProvider())
-					: KeyStore.getInstance(ssl.getKeyStoreType());
-			try {
-				URL url = ResourceUtils.getURL(ssl.getKeyStore());
-				store.load(url.openStream(),
-						ssl.getKeyStorePassword() != null ? ssl.getKeyStorePassword().toCharArray() : null);
-			}
-			catch (Exception e) {
-				throw new RuntimeException("Could not load key store ' " + ssl.getKeyStore() + "'", e);
-			}
-
-			return store;
-		}
-		catch (KeyStoreException | NoSuchProviderException e) {
-			throw new RuntimeException("Could not load KeyStore for given type and provider", e);
-		}
-	}
-
-	protected void setTrustManager(SslContextBuilder sslContextBuilder, X509Certificate... trustedX509Certificates) {
-		sslContextBuilder.trustManager(trustedX509Certificates);
-	}
-
-	protected void setTrustManager(SslContextBuilder sslContextBuilder, TrustManagerFactory factory) {
-		sslContextBuilder.trustManager(factory);
-	}
 
 	protected ConnectionProvider buildConnectionProvider(HttpClientProperties properties) {
 		HttpClientProperties.Pool pool = properties.getPool();
